Moved prefs to WPPermissionUtils
